Refactor sensors into separate binary, non-binary, and shared modules.
